reorganize db tables
